Use src-side launcher for telemetry_unittests in chromium_trybot recipe (attempt #2)
This is a re-landing of https://codereview.chromium.org/666023004/ with fixes for Windows.

Depends on https://codereview.chromium.org/677023002/

BUG=422235

Review URL: https://codereview.chromium.org/679523003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@292672 0039d316-1c4b-4281-b951-d872f2087c98

+++ recipe_modules/raw_io/api.py
@@ -34,7 +34,8 @@ class InputDataPlaceholder(recipe_util.Placeholder):
 
   def result(self, presentation, test):
     assert self._backing_file
-    if not test.enabled:  # pragma: no cover
+    exists = os.path.exists(self._backing_file)
+    if not test.enabled and exists:  # pragma: no cover
       os.unlink(self._backing_file)
     self._backing_file = None
